refactor(forge): split Forgejo access methods into forgejo_access.go (#1025)
